Refactor: #101 - privateRoute 분리
app.tsx에서 privateRoute를 분리해서 가독성을 높임
