+++ client/src/App.tsx
@@ -2,7 +2,6 @@
 import {
   NotFoundPage,
   MainPage,
-  ReviewPage,
   ReviewSubmitPage,
   LoadPage,
   RankingPage,
@@ -17,19 +16,11 @@ import GlobalStyle from '@styledComponents/GlobalStyle';
 import myTheme from '@styledComponents/theme';
 import Header from '@components/Header/index';
 import Snackbar from '@components/Snackbar';
-import { isJwtExpired } from 'jwt-check-expiration';
+import PrivateRoute from '@routes/PrivateRoute';
 
-import React, { useState, useCallback, useEffect } from 'react';
-import {
-  Switch,
-  Route,
-  Redirect,
-  useLocation,
-  RouterProps,
-} from 'react-router-dom';
+import React from 'react';
+import { Switch, Route, Redirect } from 'react-router-dom';
 import styled, { ThemeProvider } from 'styled-components';
-import { useRecoilState } from 'recoil';
-import { authState } from '@stores/atoms';
 
 const ContentWrapper = styled.div`
   width: 100%;
@@ -38,42 +29,6 @@ const ContentWrapper = styled.div`
   flex-direction: column;
   align-items: center;
 `;
-
-const PrivateRoute: React.FC<RouterProps> = ({
-  component: Component,
-  ...rest
-}) => {
-  const [auth, setAuth] = useRecoilState(authState);
-  const location = useLocation();
-
-  const checkAll = (props) => {
-    try {
-      const token = sessionStorage.getItem('jwt');
-
-      if (!token) {
-        //로그인을 안 한 상태
-        return <Redirect to={{ pathname: '/map/signin' }} />;
-      }
-      if (!auth.isLoggedin) {
-        //로그인은 했지만 새로고침
-        return <Redirect to={{ pathname: '/loading', state: location }} />;
-      }
-
-      if (isJwtExpired(token)) {
-        //로그인한 상태인데 token이 만료
-        return <Redirect to={{ pathname: '/loading', state: location }} />;
-      } else {
-        //로그인한 상태이며 token이 유효한 상태
-        return <Component {...props} />;
-      }
-    } catch (error) {
-      alert(error);
-      return <Redirect to={{ pathname: '/map/signin' }} />;
-    }
-  };
-
-  return <Route {...rest} render={checkAll} />;
-};
 
 const App: React.FC = () => {
   return (
